Build the HomeSection and DetectSection
Build a HomeSection component that includes hero title
and description. Build a DetectSection component that
includes an input to enter the image URL and a button
to click once URL is added.

In App.jsx, wrap the HomeSection and DetectSection in
a div so that they could be centered collectively.