Create the /profile/:id route and update signin
In order for users to have dynamic interfaces that presents their
data, create a "/profile/:id" route that, for now, is going to return
user data, if the id matches the user id.

Also, update the logic of loop in signin route so that it return
user correctly. Before this, the loop return an error even if the
user credentials were correct and user was present in the database.